server/requirements.txt
Django>=5.2
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
gunicorn
python-dotenv
